feat: make secure-log its own subpath export
