+++ README.md
@@ -28,10 +28,20 @@ npm i pino-roll
 
 You need an API route that will receive all the log statements.
 
+**For app dir:**
+
+Create a new API route `/app/api/logger/route.ts`, it should look like this:
+
+```ts
+export { POST } from '@navikt/next-logger/app-dir'
+```
+
+**For pages dir:**
+
 Create a new API route `/pages/api/logger.ts`, it should look like this:
 
 ```ts
-export { pinoLoggingRoute as default } from '@navikt/next-logger';
+export { pinoLoggingRoute as default } from '@navikt/next-logger/pages';
 ```
 
 ### Step 2: Logging
@@ -75,6 +85,24 @@ module.exports = {
 
 ## Configuration
 
+### App Dir
+
+You want this configuration to execute as early as possible, but on the actual client. Typically in your app-dir app, you will have for example a `<Providers>`-client that is `"use client"`-enabled.
+
+On the _root_ of any `"use client"`-enabled file that wraps your root layout.tsx, you can configure the library, for example:
+
+```ts
+"use client"
+
+configureLogger({
+    basePath: '/my/base/path',
+})
+
+export const MyProviders() {
+   ...
+}
+```
+
 ### Pages
 
 If your application is using a base path, or you want to have your logger on a different API-route, you can configure the logger.
@@ -96,27 +124,50 @@ configureLogger({
 })
 ```
 
-### App Dir
-
-You want this configuration to execute as early as possible, but on the actual client. Typically in your app-dir app, you will have for example a `<Providers>`-client that is `"use client"`-enabled.
-
-On the _root_ of any `"use client"`-enabled file that wraps your root layout.tsx, you can configure the library, for example:
-
-```ts
-"use client"
-
-configureLogger({
-    basePath: '/my/base/path',
-})
-
-export const MyProviders() {
-   ...
-}
-```
-
 ## Securelogs
 
-If you want to log sensitive information, you can use the `securelogger` function. This will instead of logging to stdout log to a file on /secure-logs. 
-This requires some setup, see [nais doc](https://doc.nais.io/observability/logging/how-to/enable-secure-logs/) for how to enable secure logging in your app.
+If you want to log sensitive information, you can use the `secureLogger` function. This will instead of logging to stdout log to a file on /secure-logs. 
+This requires some setup, see [nais docs](https://doc.nais.io/observability/logging/how-to/enable-secure-logs/) for how to enable secure logging in your app.
 
 The log file is setup with [pino-roll](https://www.npmjs.com/package/pino-roll) for rolling the logs based on file size.
+
+Using secure logger as an isomorphic logger requires an additonal API-route in your next app, the configuration is similar to the primary logger route.
+
+**For app dir:**
+
+Create a new API route `/app/api/secure-logger/route.ts`, it should look like this:
+
+```ts
+export { POST } from '@navikt/next-logger/secure-log/app-dir'
+```
+
+**For pages dir:**
+
+Create a new API route `/pages/api/secure-logger.ts`, it should look like this:
+
+```ts
+export { pinoLoggingRoute as default } from '@navikt/next-logger/secure-log/pages';
+```
+
+## Breaking changes: migrating to v2
+
+The only breaking change is that the paths for the API routes have been updated.
+
+### App Dir
+
+`app/api/logger/route.ts`:
+
+```diff
+- export { POSTLoggingRouteHandler as POST } from '@navikt/next-logger'
++ export { POST } from '@navikt/next-logger/app-dir'
+```
+### Pages
+
+`pages/api/logger/route.ts`:
+
+```diff
+- export { pinoLoggingRoute as default } from '@navikt/next-logger'
++ export { pinoLoggingRoute as default } from '@navikt/next-logger/pages'
+```
+
+If you want to use the new secureLogger feature, refer to the Securelogs docs above.
