+++ src/index.ts
@@ -1,5 +1,4 @@
 export * from './logger'
-export * from './loggingRoute'
-export * from './loggingRouteHandler'
-export { logger as backendLogger } from './backendLogger'
+
+export { backendLogger } from './loggers/backendLogger'
 export { configureLogger } from './config'
