+++ src/logger.ts
@@ -1,11 +1,9 @@
 import pino from 'pino'
 
-import { logger as backendLogger } from './backendLogger'
-import { logger as frontendLogger } from './frontendLogger'
+import { backendLogger as backendLogger } from './loggers/backendLogger'
+import { frontendLogger as frontendLogger } from './loggers/frontendLogger'
 
 export const logger = typeof window !== 'undefined' ? frontendLogger() : backendLogger()
-
-export const secureLogger = typeof window !== 'undefined' ? frontendLogger(true) : backendLogger({ secure: true })
 
 export const createChildLogger = (requestId: string): pino.Logger => {
     return logger.child({ x_request_id: requestId })
