+++ src/routes/app-dir/index.ts
@@ -1,0 +1,5 @@
+import { backendLogger } from '../../loggers/backendLogger'
+
+import { createLoggingRouteHandler } from './loggingRouteHandler'
+
+export const POST = createLoggingRouteHandler(backendLogger())
