+++ src/routes/pages/index.ts
@@ -1,0 +1,5 @@
+import { backendLogger } from '../../loggers/backendLogger'
+
+import { createLoggingApiRoute } from './loggingApiRoute'
+
+export const pinoLoggingRoute = createLoggingApiRoute(backendLogger())
