+++ src/secure-log/index.ts
@@ -1,0 +1,3 @@
+export * from './logger'
+
+export { secureBackendLogger } from './secureBackendLogger'
