+++ src/secure-log/logger.ts
@@ -1,0 +1,4 @@
+import { secureBackendLogger } from './secureBackendLogger'
+import { frontendLogger as frontendLogger } from '../loggers/frontendLogger'
+
+export const secureLogger = typeof window !== 'undefined' ? frontendLogger(true) : secureBackendLogger()
