+++ src/secure-log/routes/app-dir/index.ts
@@ -1,0 +1,4 @@
+import { createLoggingRouteHandler } from '../../../routes/app-dir/loggingRouteHandler'
+import { secureBackendLogger } from '../../secureBackendLogger'
+
+export const POST = createLoggingRouteHandler(secureBackendLogger())
