+++ src/secure-log/routes/pages/index.ts
@@ -1,0 +1,4 @@
+import { createLoggingApiRoute } from '../../../routes/pages/loggingApiRoute'
+import { secureBackendLogger } from '../../secureBackendLogger'
+
+export const pinoLoggingRoute = createLoggingApiRoute(secureBackendLogger())
